Move `StorageEngine` class to a new module and update imports
* **src/tinychain/tinychain.py**
  - Remove `StorageEngine` class
  - Import `StorageEngine` class from new `storage_engine` module

* **src/tinychain/storage_engine.py**
  - Create new module for `StorageEngine` class
  - Move `StorageEngine` class from `tinychain.py` to this new module
  - Import necessary dependencies
